% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLocationValidity.R
\name{checkLocationValidity}
\alias{checkLocationValidity}
\title{Check the validity of a location}
\usage{
checkLocationValidity(location)
}
\arguments{
\item{location}{a string representing the desired location}
}
\value{
an ERROR if the location is not in the yrbss_data_binary dataset,
the location itself if it is valid, and "US" if it is NULL.
}
\description{
Helper method to check the validity of a location.
}
\examples{
checkLocationValidity("CA")

}
