% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVariableValidity.R
\name{checkVariableValidity}
\alias{checkVariableValidity}
\title{Check the validity of a variable name}
\usage{
checkVariableValidity(variable)
}
\arguments{
\item{variable}{a string representing the name of the desired binary variable}
}
\value{
an ERROR if the variable is NULL or not in the yrbss_questions_binary
list of questions
}
\description{
Helper method to check the validity of a variable name.
}
\examples{
checkVariableValidity("qn8")

}
