% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkYearValidity.R
\name{checkYearValidity}
\alias{checkYearValidity}
\title{Check the validity of a year}
\usage{
checkYearValidity(year)
}
\arguments{
\item{year}{a four-digit number representing the desired year}
}
\value{
an ERROR if the year is not in the yrbss_data_binary dataset,
the year itself if it is valid, and 2015 if it is NULL.
}
\description{
Helper method to check the validity of a year
}
\examples{
checkLocationValidity("CA")

}
