% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListOfParticipatingStates.R
\name{getListOfParticipatingStates}
\alias{getListOfParticipatingStates}
\title{YRBSS Data - Binary Version - List of Participating States}
\usage{
getListOfParticipatingStates()
}
\value{
vector of participating US states abbreviations
}
\description{
Helper method to extract the list of participating US states in the
\code{\link{yrbss_states_data_binary}} data.
}
\examples{
getListOfParticipatingStates()

}
