% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\docType{data}
\name{survey}
\alias{survey}
\title{Youth Risk Behavior Surveillance System}
\format{A data frame with 173,274 observations and 231 variables.}
\usage{
survey
}
\description{
The Youth Risk Behavior Surveillance System (YRBSS) monitors six types of
health-risk behaviors that contribute to the leading causes of death and
disability among youth and adults, including:
\itemize{
\item Behaviors that contribute to unintentional injuries and violence
\item Sexual behaviors that contribute to unintended pregnancy and
  sexually transmitted diseases, including HIV infection
\item Alcohol and other drug use
\item Tobacco use
\item Unhealthy dietary behaviors
\item Inadequate physical activity
}
}
\details{
The data includes a very large number of variables, so please consult
the bundled \code{system.file("documentation.pdf", package = "yrbss")}
for details.
}
\keyword{datasets}

