% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportionSingleVariable.R
\name{getProportionSingleVariable}
\alias{getProportionSingleVariable}
\title{Get a Single Variable "Yes" Proportion}
\usage{
getProportionSingleVariable(variable = NULL, location = "US",
  .year = NULL, label = NULL, filters = NULL, .level = 0.95)
}
\arguments{
\item{variable}{a string representing the name of the desired binary variable (if not specified "label" must be specified)}

\item{location}{a string representing the desired location, defaults to "US"}

\item{.year}{a int of the desired year (currently: 1991 to 2015)}

\item{label}{optionally instead of a "variable" argument one could specify the variable label}

\item{filters}{a vector of strings representing desired filters (see details)}

\item{.level}{the desired confidence interval level, defaults to 0.95}
}
\value{
a list containing:
\item{prop}{the proportion of "Yes" for the desired variables within given filters}
\item{ciLB}{Confidence Interval lower bound}
\item{ciUB}{Confidence Interval upper bound}
}
\description{
Retrieves the proportion and confidence interval for a single variable in the
binary version, for given location and year as well as additional filters
}
\examples{
getProportionSingleVariable("qn8", "US", 2015)

}
