% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yrbss_data_binary.R
\docType{data}
\name{yrbss_data_binary}
\alias{yrbss_data_binary}
\title{YRBSS Data - Binary Version}
\format{An object of class \code{"tbl_df"};}
\source{
\href{www.cdc.gov/yrbs}{CDC YRBSS Official Site}
}
\usage{
data(yrbss_data)
}
\description{
Data from CDC YRBSS - the Binary Version, where multiple choice questions were grouped into two options.
}
\examples{
data(yrbss_data_binary)
\donttest{iplotCurves(phe, times)}
}
\references{
Centers for Disease Control and Prevention. Youth Risk Behavior Survey Questionnaire. Accessed on [2017-06-14].
(\href{www.cdc.gov/yrbs}{CDC YRBSS Official Site})
}
\keyword{datasets}
